move to util class
